Interpolation of method change: from lagrange to cubic (with the possibility on making any kind of interpolation in wavelength_interpolation)
The objective is to make the preprocessing of the spectra much faster, and more robust.

Lagrange interpolation is a wiser solution when there is only a few spectrum. Linear interpolation is better for high througput application at the time being.

With a much faster Lagrange interpolation method than my solution  (that is still available in option), I could change my jugement.

git-svn-id: https://scm.gforge.inria.fr/authscm/pradal/svn/openaleapkg/trunk@1251 0978ab36-7443-0410-8248-c2aa344a765a